initial upload
for using these files, take Syftmax and place these files in the folder. main.cpp, progression.cpp, and progression.h should be in src-folder. all examples should be in a new map called 'example'. the Python domain generators are not used by Syftmax, but only for generating the LTLf specifications as written in the examples (can be replaced if wished)